Rename layers module to layerModel, remove original
